Logic code for select elements and showing result based on choices
Populated Select element with all possible options for the specific categories.

Created object which contins keys, values are assigned to appropriate keys once
search button is clicked.

created method which takes those keys and values and inserts them into the url
to return the appropriate location of the crimes in accordance the users choice.