Addition of Promises, JQuery AJAX, refractoring of code
Global variables were declared so the the map object can be referenced and updates through
out the project.

loadMap is a promise that is called on window.onload, it will load an empty map and
serves as a good background. The map will not be loaded repeatedly every time the user clicks
the search button.

Using JQuery Ajax which returns the $.ajax object (which is a promise) and calling the .done method
and passing in the data to initialize the map other wise .fail method will be console.log the errors.